tests refactor: split test file into smaller ones
